Added pre-game screen and refactored naming. See description
- Renamed 'LobbyScreen' to 'DeploymentScreen'
- Added new 'LobbyScreen' that takes player name, reports time until game start
- Two new connectionReporter states, 'GameWaitingToStart' and 'GameReady'
- ConnectionReporter value 'TimeUntilGameStart' available
