+++ workers/unity/Assets/Fps/Scripts/UI/ConnectButton/ConnectionStatusUIController.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections;
 using UnityEngine;
 using UnityEngine.UI;
 
@@ -16,6 +17,10 @@ namespace Fps
         public string FailedToGetDeploymentListText = "Failed to get deployment list!";
         public string ConnectingText = "Searching for deployment to join...";
         public string ConnectionFailedText = "Failed to join deployment!";
+        public string ConnectedText = "Connected";
+        public string WaitingForGameStartText = "Game starting in {0:#} SECONDS";
+        public string WaitingForGameToStartText_NoNumber = "Game starting...";
+        public string GameReadyText = "Game begun! Ready to spawn.";
         public string SpawningText = "Joining deployment...";
         public string SpawningFailedText = "Failed to spawn player!";
         public string WorkerDisconnectedText = "Worker was disconnected";
@@ -55,10 +60,16 @@ namespace Fps
                     State_Connecting();
                     break;
                 case ConnectionStateReporter.State.Connected:
-                    ConnectionStateReporter.TrySpawn();
+                    State_Connected();
                     break;
                 case ConnectionStateReporter.State.ConnectionFailed:
                     State_ConnectionFailed(information);
+                    break;
+                case ConnectionStateReporter.State.WaitingForGameStart:
+                    State_WaitingForGameStart();
+                    break;
+                case ConnectionStateReporter.State.GameReady:
+                    State_GameReady();
                     break;
                 case ConnectionStateReporter.State.Spawning:
                     State_Spawning();
@@ -106,10 +117,28 @@ namespace Fps
             SetSymbol(SpinnerSymbol);
         }
 
+        private void State_Connected()
+        {
+            StatusText.text = ConnectedText;
+            SetSymbol(SuccessSymbol);
+        }
+
         private void State_ConnectionFailed(string error)
         {
             StatusText.text = $"{ConnectionFailedText}\nError: {error}";
             SetSymbol(ErrorSymbol);
+        }
+
+        private void State_WaitingForGameStart()
+        {
+            SetSymbol(SpinnerSymbol);
+            StartCoroutine(nameof(UpdateGameStartTime));
+        }
+
+        private void State_GameReady()
+        {
+            SetSymbol(SuccessSymbol);
+            StatusText.text = GameReadyText;
         }
 
         private void State_Spawning()
@@ -149,6 +178,19 @@ namespace Fps
             if (SuccessSymbol != null)
             {
                 SuccessSymbol.SetActive(symbol == SuccessSymbol);
+            }
+        }
+
+        private IEnumerator UpdateGameStartTime()
+        {
+            while (ConnectionStateReporter.CurrentState == ConnectionStateReporter.State.WaitingForGameStart)
+            {
+                var timeUntilStart = ConnectionStateReporter.TimeUntilGameStart;
+                StatusText.text = timeUntilStart < 1
+                    ? WaitingForGameToStartText_NoNumber
+                    : string.Format(WaitingForGameStartText, ConnectionStateReporter.TimeUntilGameStart);
+
+                yield return new WaitForSeconds(.5f);
             }
         }
     }
