+++ workers/unity/Assets/Fps/Scripts/UI/Controllers/DeploymentListScreenController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 30e847689ddbdb04cb797d4e890564db
+guid: 1cc898fb93634574a9485403c278d570
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
