+++ workers/unity/Assets/Fps/Scripts/UI/Controllers/LobbyScreenController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 1cc898fb93634574a9485403c278d570
+guid: bf303b9588e50d349a7a4b2ff52b4a73
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
