+++ workers/unity/Assets/Fps/Scripts/UI/Controllers/Testers/DeploymentListScreenTester.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 30e847689ddbdb04cb797d4e890564db
+guid: c77f5b0a978dd794a8acb7f3deae3f5b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
